IPS-560: Stage 2: Canary deployment on F2F Frontend changes
Canary deployment for ECS has 2 stages of deployment and should be deployed in stgaes.
This is deployment stage 1 where additional target group is created for green deployment,
configured to code deploy to deploy the blue green deployment. Additionally the ecs canary
stack is deployed which in turn deploys a nested stack for canary deployment. Also, added
alarms for rollback on canary deployments.